Move generic color space related stuff out into a new module.
