+++ src/channel.rs
@@ -10,7 +10,7 @@ use serde::Deserialize;
 use crate::{
     animation::AnimationUIState,
     control::EmitControlMessage,
-    fixture::{color::HsluvRenderer, FixtureGroup, FixtureGroupKey, Patch},
+    fixture::{FixtureGroup, FixtureGroupKey, Patch},
     osc::{EmitOscMessage, GroupControlMap, OscControlMessage, ScopedControlEmitter},
     wled::EmitWledControlMessage,
 };
