+++ src/show.rs
@@ -4,9 +4,10 @@ use crate::{
     animation::AnimationUIState,
     channel::{ChannelStateEmitter, Channels},
     clock_service::ClockService,
+    color::HsluvRenderer,
     control::{ControlMessage, Controller},
     dmx::DmxBuffer,
-    fixture::{color::HsluvRenderer, Patch},
+    fixture::Patch,
     master::MasterControls,
     midi::{MidiControlMessage, MidiHandler},
     osc::{GroupControlMap, OscControlMessage, ScopedControlEmitter},
